Add support for DAGMan workflows.
DAGMan[0] organizes a user's jobs into a graph of nodes. farmout can be run
several times with the new --output-dag-file and --last-submit-dir options in
order to build the graph. DAGMan will then manage the jobs in the graph,
submitting (and retrying) them as necessary. After the DAG terminates, a
notification email will be sent to the user.

While this approach has clear advantages for multi-step workflows, the
notification feature may also be useful when submitting a single batch of jobs.

Thanks to Jeff Klukas <jklukas@hep.wisc.edu> for testing this feature!

The full development history of this patch can be found in
/wcmaier/farmout-dag[1].


[0] http://www.cs.wisc.edu/condor/manual/v7.6/2_10DAGMan_Applications.html
[1] http://hg.hep.wisc.edu/wcmaier/farmout-dag/rev/dag